pyproject.toml
[tool.ruff]
exclude = [
    ".venv",
    ".git",
    "__pycache__",
    "build",
    "dist",
    "venv",
]
lint.ignore = []
line-length = 120
lint.select = [
    "E",
    "F",
    "W",
    "I001",
]
src = ["./"]

[tool.ruff.format]
quote-style = "double"

[tool.poetry]
name = "tiktokshop-be"
version = "0.1.0"
description = ""
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
asgiref = "3.7.2"
attrs = "23.1.0"
black = "23.11.0"
cachetools = "5.3.3"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
coloredlogs = "15.0.1"
crispy-bootstrap5 = "2023.10"
django = "4.2.7"
django-cors-headers = "4.3.1"
django-crispy-forms = "2.1"
django-debug-toolbar = "4.2.0"
django-widget-tweaks = "1.5.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.1"
drf-spectacular = "0.26.5"
et-xmlfile = "1.1.0"
flatbuffers = "24.3.25"
frozenlist = "1.4.1"
google-api-core = "2.19.0"
google-api-python-client = "2.129.0"
google-auth = "2.29.0"
google-auth-httplib2 = "0.2.0"
google-auth-oauthlib = "1.2.0"
googleapis-common-protos = "1.63.0"
httplib2 = "0.22.0"
humanfriendly = "10.0"
idna = "3.7"
imageio = "2.34.1"
inflection = "0.5.1"
jsonschema = "4.20.0"
jsonschema-specifications = "2023.11.2"
lazy-loader = "0.4"
llvmlite = "0.42.0"
mpmath = "1.3.0"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
networkx = "3.3"
numba = "0.59.1"
numpy = "1.26.4"
oauthlib = "3.2.2"
onnxruntime = "1.18.0"
opencv-python-headless = "4.9.0.80"
openpyxl = "3.1.2"
packaging = "23.2"
pandas = "2.2.2"
pathspec = "0.11.2"
pdf2image = "1.17.0"
pillow = "10.3.0"
platformdirs = "4.1.0"
pooch = "1.8.1"
proto-plus = "1.23.0"
protobuf = "4.25.3"
psycopg2-binary = "2.9.9"
pyasn1 = "0.6.0"
pyasn1-modules = "0.4.0"
pyjwt = "2.8.0"
pymatting = "1.1.12"
pyngrok = "7.1.6"
pyparsing = "3.1.2"
pypdf2 = "1.26.0"
pyreadline3 = "3.4.1"
pytesseract = "0.3.10"
python-dateutil = "2.9.0.post0"
python-decouple = "3.8"
python-dotenv = "1.0.1"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
referencing = "0.32.0"
rembg = "2.0.56"
requests = "2.31.0"
requests-oauthlib = "2.0.0"
rpds-py = "0.13.2"
rsa = "4.9"
scikit-image = "0.23.2"
scipy = "1.13.0"
six = "1.16.0"
sqlparse = "0.4.4"
sympy = "1.12"
tifffile = "2024.5.10"
tqdm = "4.66.4"
tzdata = "2023.3"
uritemplate = "4.1.1"
urllib3 = "2.2.1"
yarl = "1.9.4"
isort = "^5.13.2"
django-filter = "^24.2"
httpx = "^0.27.0"
